refactor expr
